# feat(bloomstore): Introduce specialised cache for blocks