--- a/pkg/storage/stores/shipper/bloomshipper/blockscache.go
+++ b/pkg/storage/stores/shipper/bloomshipper/blockscache.go
@@ -1,0 +1,490 @@
+package bloomshipper
+
+import (
+	"container/list"
+	"context"
+	"flag"
+	"fmt"
+	"os"
+	"sync"
+	"time"
+
+	"github.com/go-kit/log"
+	"github.com/go-kit/log/level"
+	"github.com/grafana/dskit/flagext"
+	"github.com/pkg/errors"
+	"github.com/prometheus/client_golang/prometheus"
+	"github.com/prometheus/client_golang/prometheus/promauto"
+	"go.uber.org/atomic"
+
+	"github.com/grafana/loki/pkg/util"
+	"github.com/grafana/loki/pkg/util/constants"
+)
+
+const (
+	defaultPurgeInterval = 1 * time.Minute
+
+	// eviction reasons
+	reasonExpired = "expired"
+	reasonFull    = "full"
+
+	// errors when putting entries
+	errAlreadyExists = "entry already exists: %s"
+	errTooBig        = "entry exceeds hard limit: %s"
+)
+
+type Cache interface {
+	Put(ctx context.Context, key string, value BlockDirectory) error
+	PutInc(ctx context.Context, key string, value BlockDirectory) error
+	PutMany(ctx context.Context, keys []string, values []BlockDirectory) error
+	Get(ctx context.Context, key string) (BlockDirectory, bool)
+	Release(ctx context.Context, key string) error
+	Stop()
+}
+
+type blocksCacheMetrics struct {
+	entriesAdded   prometheus.Counter
+	entriesEvicted *prometheus.CounterVec
+	entriesFetched *prometheus.CounterVec
+	entriesCurrent prometheus.Gauge
+	usageBytes     prometheus.Gauge
+}
+
+func newBlocksCacheMetrics(reg prometheus.Registerer, namespace, subsystem string) *blocksCacheMetrics {
+	return &blocksCacheMetrics{
+		entriesAdded: promauto.With(reg).NewCounter(prometheus.CounterOpts{
+			Namespace: namespace,
+			Subsystem: subsystem,
+			Name:      "added_total",
+			Help:      "The total number of entries added to the cache",
+		}),
+		entriesEvicted: promauto.With(reg).NewCounterVec(prometheus.CounterOpts{
+			Namespace: namespace,
+			Subsystem: subsystem,
+			Name:      "evicted_total",
+			Help:      "The total number of entries evicted from the cache",
+		}, []string{"reason"}),
+		entriesFetched: promauto.With(reg).NewCounterVec(prometheus.CounterOpts{
+			Namespace: namespace,
+			Subsystem: subsystem,
+			Name:      "fetched_total",
+			Help:      "Total number of entries fetched from cache, grouped by status",
+		}, []string{"status"}),
+		entriesCurrent: promauto.With(reg).NewGauge(prometheus.GaugeOpts{
+			Namespace: namespace,
+			Subsystem: subsystem,
+			Name:      "entries",
+			Help:      "Current number of entries in the cache",
+		}),
+		usageBytes: promauto.With(reg).NewGauge(prometheus.GaugeOpts{
+			Namespace: namespace,
+			Subsystem: subsystem,
+			Name:      "usage_bytes",
+			Help:      "The current size of entries managed by the cache in bytes",
+		}),
+	}
+}
+
+// compiler check to enforce implementation of the Cache interface
+var _ Cache = &BlocksCache{}
+
+// BlocksCache is an in-memory cache that manages block directories on the filesystem.
+type BlocksCache struct {
+	cfg     BlocksCacheConfig
+	metrics *blocksCacheMetrics
+	logger  log.Logger
+
+	lock    sync.RWMutex // lock for cache entries
+	entries map[string]*list.Element
+	lru     *list.List
+
+	done chan struct{}
+
+	cond        *sync.Cond // condition for soft eviction
+	shouldEvict bool
+
+	currSizeBytes int64
+}
+
+type Entry struct {
+	Key   string
+	Value BlockDirectory
+
+	created  time.Time
+	accessed time.Time
+
+	refCount *atomic.Int32
+}
+
+// BlocksCacheConfig represents in-process embedded cache config.
+type BlocksCacheConfig struct {
+	Enabled   bool          `yaml:"enabled,omitempty"`
+	SoftLimit flagext.Bytes `yaml:"soft_limit"`
+	HardLimit flagext.Bytes `yaml:"hard_limit"`
+	TTL       time.Duration `yaml:"ttl"`
+
+	// PurgeInterval tell how often should we remove keys that are expired.
+	// by default it takes `defaultPurgeInterval`
+	PurgeInterval time.Duration `yaml:"-"`
+}
+
+func (cfg *BlocksCacheConfig) RegisterFlagsWithPrefix(prefix, description string, f *flag.FlagSet) {
+	cfg.RegisterFlagsWithPrefixAndDefaults(prefix, description, f, time.Hour)
+}
+
+func (cfg *BlocksCacheConfig) RegisterFlagsWithPrefixAndDefaults(prefix, description string, f *flag.FlagSet, defaultTTL time.Duration) {
+	f.BoolVar(&cfg.Enabled, prefix+"enabled", false, description+"Whether blocks cache is enabled.")
+	f.Var(&cfg.SoftLimit, prefix+"soft-limit", description+"Soft limit of the cache in bytes. Exceeding this limit will trigger evictions of least recently used items in the background.")
+	_ = cfg.SoftLimit.Set("32GiB")
+	f.Var(&cfg.HardLimit, prefix+"hard-limit", description+"Hard limit of the cache in bytes. Exceeding this limit will block execution until soft limit is deceeded.")
+	_ = cfg.HardLimit.Set("64GiB")
+	f.DurationVar(&cfg.TTL, prefix+"ttl", defaultTTL, description+"The time to live for items in the cache before they get purged.")
+}
+
+func (cfg *BlocksCacheConfig) IsEnabled() bool {
+	return cfg.Enabled
+}
+
+func (cfg *BlocksCacheConfig) Validate() error {
+	if !cfg.Enabled {
+		return nil
+	}
+	if cfg.TTL == 0 {
+		return errors.New("blocks cache ttl must not be 0")
+	}
+	if cfg.SoftLimit == 0 {
+		return errors.New("blocks cache soft_limit must not be 0")
+	}
+	if cfg.SoftLimit > cfg.HardLimit {
+		return errors.New("blocks cache soft_limit must not be greater than hard_limit")
+	}
+	return nil
+}
+
+// NewFsBlocksCache returns a new initialised BlocksCache with the key and value of requested types.
+// To limit the memory allocated by items in the cache, it's necessary to pass cacheEntrySizeCalculator
+// that calculates the size of an entry in bytes.
+// Also, this constructor allows passing the callback that will be called for the entry whenever it is removed from the cache.
+func NewFsBlocksCache(cfg BlocksCacheConfig, reg prometheus.Registerer, logger log.Logger) *BlocksCache {
+	cache := &BlocksCache{
+		cfg:     cfg,
+		logger:  logger,
+		metrics: newBlocksCacheMetrics(reg, constants.Loki, "bloom_blocks_cache"),
+		entries: make(map[string]*list.Element),
+		lru:     list.New(),
+		done:    make(chan struct{}),
+		cond:    sync.NewCond(&sync.Mutex{}),
+	}
+
+	// Set a default interval for the ticker
+	// This can be overwritten to a smaller value in tests
+	if cfg.PurgeInterval == 0 {
+		cfg.PurgeInterval = defaultPurgeInterval
+	}
+
+	go cache.runTTLEvictJob(cfg.PurgeInterval, cfg.TTL)
+	go cache.runLRUevictJob()
+
+	return cache
+}
+
+// Put implements Cache.
+// It stores a value with given key.
+func (c *BlocksCache) Put(ctx context.Context, key string, value BlockDirectory) error {
+	if ctx.Err() != nil {
+		return ctx.Err()
+	}
+
+	c.lock.Lock()
+	defer c.lock.Unlock()
+	_, err := c.put(key, value)
+	return err
+}
+
+// PutInc implements Cache.
+// It stores a value with given key and increments the ref counter on that item.
+func (c *BlocksCache) PutInc(ctx context.Context, key string, value BlockDirectory) error {
+	if ctx.Err() != nil {
+		return ctx.Err()
+	}
+
+	c.lock.Lock()
+	defer c.lock.Unlock()
+
+	entry, err := c.put(key, value)
+	if err != nil {
+		return err
+	}
+
+	entry.accessed = time.Now()
+	entry.refCount.Inc()
+	return nil
+}
+
+// PutMany implements Cache.
+func (c *BlocksCache) PutMany(ctx context.Context, keys []string, values []BlockDirectory) error {
+	if ctx.Err() != nil {
+		return ctx.Err()
+	}
+
+	c.lock.Lock()
+	defer c.lock.Unlock()
+
+	var err util.MultiError
+	for i := range keys {
+		if _, e := c.put(keys[i], values[i]); e != nil {
+			err.Add(e)
+		}
+	}
+	return err.Err()
+}
+
+func (c *BlocksCache) put(key string, value BlockDirectory) (*Entry, error) {
+	// Blocks cache does not allow updating, so it rejects values with the same key
+	_, exists := c.entries[key]
+	if exists {
+		return nil, fmt.Errorf(errAlreadyExists, key)
+	}
+
+	entry := &Entry{
+		Key:      key,
+		Value:    value,
+		created:  time.Now(),
+		refCount: atomic.NewInt32(0),
+	}
+	size := c.sizeOf(entry)
+
+	// Reject items that are larger than the hard limit
+	if size > int64(c.cfg.HardLimit) {
+		// It's safe to clean up the disk, since it does not have any references
+		// yet. Ideally, we avoid downloading blocks that do not fit into the cache
+		// upfront.
+		_ = c.remove(entry)
+		return nil, fmt.Errorf(errTooBig, key)
+	}
+
+	// Allow adding the new item even if the cache exceeds its soft limit.
+	// However, broadcast the condition that the cache should be cleaned up.
+	if c.currSizeBytes+size > int64(c.cfg.SoftLimit) {
+		level.Debug(c.logger).Log(
+			"msg", "adding item exceeds soft limit",
+			"action", "trigger soft eviction",
+			"curr_size_bytes", c.currSizeBytes,
+			"entry_size_bytes", size,
+			"soft_limit_bytes", c.cfg.SoftLimit,
+			"hard_limit_bytes", c.cfg.HardLimit,
+		)
+		c.cond.L.Lock()
+		c.shouldEvict = true
+		c.cond.L.Unlock()
+
+		c.cond.Signal()
+	}
+
+	// Adding an item blocks if the cache would exceed its hard limit.
+	if c.currSizeBytes+size > int64(c.cfg.HardLimit) {
+		level.Debug(c.logger).Log(
+			"msg", "adding item exceeds hard limit",
+			"action", "evict items until space is freed",
+			"curr_size_bytes", c.currSizeBytes,
+			"entry_size_bytes", size,
+			"soft_limit_bytes", c.cfg.SoftLimit,
+			"hard_limit_bytes", c.cfg.HardLimit,
+		)
+		// TODO(chaudum): implement case
+		return nil, errors.New("todo: implement waiting for evictions to free up space")
+	}
+
+	// Cache has space to add the item.
+	c.entries[key] = c.lru.PushFront(entry)
+	c.currSizeBytes += size
+
+	c.metrics.entriesAdded.Inc()
+	c.metrics.entriesCurrent.Inc()
+	c.metrics.usageBytes.Set(float64(c.currSizeBytes))
+	return entry, nil
+}
+
+func (c *BlocksCache) evict(key string, element *list.Element, reason string) {
+	entry := element.Value.(*Entry)
+	err := c.remove(entry)
+	if err != nil {
+		level.Error(c.logger).Log("msg", "failed to remove entry from disk", "err", err)
+		return
+	}
+	c.lru.Remove(element)
+	delete(c.entries, key)
+	c.currSizeBytes -= c.sizeOf(entry)
+	c.metrics.entriesCurrent.Dec()
+	c.metrics.entriesEvicted.WithLabelValues(reason).Inc()
+}
+
+// Get implements Cache.
+// Get returns the stored value against the given key.
+func (c *BlocksCache) Get(ctx context.Context, key string) (BlockDirectory, bool) {
+	if ctx.Err() != nil {
+		return BlockDirectory{}, false
+	}
+
+	c.lock.RLock()
+	defer c.lock.RUnlock()
+
+	entry := c.get(ctx, key)
+	if entry == nil {
+		return BlockDirectory{}, false
+	}
+	return entry.Value, true
+}
+
+func (c *BlocksCache) get(_ context.Context, key string) *Entry {
+	element, exists := c.entries[key]
+	if !exists {
+		c.metrics.entriesFetched.WithLabelValues("miss").Inc()
+		return nil
+	}
+
+	entry := element.Value.(*Entry)
+	entry.accessed = time.Now()
+	entry.refCount.Inc()
+
+	c.lru.MoveToFront(element)
+
+	c.metrics.entriesFetched.WithLabelValues("hit").Inc()
+	return entry
+}
+
+// Release decrements the ref counter on the cached item with given key.
+func (c *BlocksCache) Release(ctx context.Context, key string) error {
+	if ctx.Err() != nil {
+		return ctx.Err()
+	}
+
+	// We can use a read lock because we only update a field on an existing entry
+	// and we do not modify the map of entries or the order in the LRU list.
+	c.lock.RLock()
+	defer c.lock.RUnlock()
+
+	element, exists := c.entries[key]
+	if !exists {
+		return nil
+	}
+
+	entry := element.Value.(*Entry)
+	entry.refCount.Dec()
+	return nil
+}
+
+// Stop implements Cache.
+func (c *BlocksCache) Stop() {
+	c.lock.Lock()
+	defer c.lock.Unlock()
+
+	c.entries = make(map[string]*list.Element)
+	c.lru.Init()
+
+	c.metrics.entriesCurrent.Set(float64(0))
+	c.metrics.usageBytes.Set(float64(0))
+
+	close(c.done)
+}
+
+func (c *BlocksCache) sizeOf(entry *Entry) int64 {
+	return entry.Value.Size()
+}
+
+func (c *BlocksCache) remove(entry *Entry) error {
+	level.Info(c.logger).Log("msg", "remove entry from disk", "path", entry.Value.Path)
+	err := os.RemoveAll(entry.Value.Path)
+	if err != nil {
+		return fmt.Errorf("error removing bloom block directory from disk: %w", err)
+	}
+	return nil
+}
+
+func (c *BlocksCache) runLRUevictJob() {
+	level.Info(c.logger).Log("msg", "run lru evict job")
+	for {
+		select {
+		case <-c.done:
+			return
+		default:
+			c.cond.L.Lock()
+			for !c.shouldEvict {
+				c.cond.Wait()
+			}
+			c.shouldEvict = false
+			c.cond.L.Unlock()
+			c.evictLeastRecentlyUsedItems()
+		}
+	}
+}
+
+func (c *BlocksCache) evictLeastRecentlyUsedItems() {
+	c.lock.Lock()
+	defer c.lock.Unlock()
+
+	level.Debug(c.logger).Log(
+		"msg", "evict least recently used entries",
+		"curr_size_bytes", c.currSizeBytes,
+		"soft_limit_bytes", c.cfg.SoftLimit,
+		"hard_limit_bytes", c.cfg.HardLimit,
+	)
+	for c.currSizeBytes >= int64(c.cfg.SoftLimit) {
+		elem := c.lru.Back()
+		if elem == nil {
+			break
+		}
+		entry := elem.Value.(*Entry)
+		if entry.refCount.Load() == 0 {
+			level.Debug(c.logger).Log(
+				"msg", "evict least recently used entry",
+				"entry", entry.Key,
+			)
+			c.evict(entry.Key, elem, reasonFull)
+		}
+	}
+}
+
+func (c *BlocksCache) runTTLEvictJob(interval, ttl time.Duration) {
+	if interval == 0 || ttl == 0 {
+		return
+	}
+
+	level.Info(c.logger).Log("msg", "run ttl evict job")
+	ticker := time.NewTicker(interval)
+	defer ticker.Stop()
+
+	for {
+		select {
+		case <-c.done:
+			return
+		case <-ticker.C:
+			c.evictExpiredItems(ttl)
+		}
+	}
+}
+
+// evictExpiredItems prunes items in the cache that exceeded their ttl after last access
+func (c *BlocksCache) evictExpiredItems(ttl time.Duration) {
+	c.lock.Lock()
+	defer c.lock.Unlock()
+
+	level.Debug(c.logger).Log(
+		"msg", "evict expired entries",
+		"curr_size_bytes", c.currSizeBytes,
+		"soft_limit_bytes", c.cfg.SoftLimit,
+		"hard_limit_bytes", c.cfg.HardLimit,
+	)
+	for k, v := range c.entries {
+		entry := v.Value.(*Entry)
+		if time.Since(entry.created) > ttl && entry.refCount.Load() == 0 {
+			level.Debug(c.logger).Log(
+				"msg", "evict expired entry",
+				"entry", entry.Key,
+				"age", time.Since(entry.created),
+				"ttl", ttl,
+			)
+			c.evict(k, v, reasonExpired)
+		}
+	}
+}