**What this PR does / why we need it**:

The blocks cache is an in-memory cache that represents the single source of truth of what blocks are available on the file system and manages their EOL lifecycle.

**Special notes for your reviewer**:

There is a follow-up branch to wire up the new cache, based on this branch. See https://github.com/grafana/loki/compare/chaudum/blocks-cache-impl?expand=1