@chaudum 2024-03-20
Regarding commit https://github.com/grafana/loki/pull/12257/commits/a1ad00df295f5449ce494d2fed620c743be8d577

Before:
```
$ gotest -v -count=1 ./pkg/storage/stores/shipper/bloomshipper/... -test.run=^$ -test.bench=Benchmark_ -benchtime=1000x
goos: linux
goarch: amd64
pkg: github.com/grafana/loki/pkg/storage/stores/shipper/bloomshipper
cpu: 11th Gen Intel(R) Core(TM) i7-1185G7 @ 3.00GHz
Benchmark_BlocksCacheOld
Benchmark_BlocksCacheOld-8   	    1000	     39545 ns/op	     679 B/op	       7 allocs/op
Benchmark_BlocksCacheNew
Benchmark_BlocksCacheNew-8   	    1000	    177112 ns/op	   32719 B/op	    2008 allocs/op
```

After:

```
$ gotest -v -count=1 ./pkg/storage/stores/shipper/bloomshipper/... -test.run=^$ -test.bench=Benchmark_ -benchtime=1000x 
goos: linux
goarch: amd64
pkg: github.com/grafana/loki/pkg/storage/stores/shipper/bloomshipper
cpu: 11th Gen Intel(R) Core(TM) i7-1185G7 @ 3.00GHz
Benchmark_BlocksCacheOld
Benchmark_BlocksCacheOld-8   	    1000	     41728 ns/op	     678 B/op	       7 allocs/op
Benchmark_BlocksCacheNew
Benchmark_BlocksCacheNew-8   	    1000	     50552 ns/op	     701 B/op	       8 allocs/op
```